+++ src/Modules/Profile/ProfileView.php
@@ -229,13 +229,17 @@ class ProfileView extends View
             }
         }
 
+        $formattedLastActivity = ($this->foodsaver['last_activity'] !== '0000-00-00 00:00:00')
+            ? Carbon::parse($this->foodsaver['last_activity'])->format('d.m.Y')
+            : null;
+
         $page->addSectionLeft(
             $this->vueComponent('vue-profile-infos', 'ProfileInfos', [
                 'isfoodsaver' => $this->foodsaver['rolle'] > Role::FOODSHARER,
                 'fsMail' => $fsMail,
                 'privateMail' => $this->profilePermissions->maySeePrivateEmail($fsId) ? $this->foodsaver['email'] : '',
                 'registrationDate' => $this->profilePermissions->maySeeRegistrationDate($fsId) ? Carbon::parse($this->foodsaver['anmeldedatum'])->format('d.m.Y') : '',
-                'lastActivity' => $this->profilePermissions->maySeelastActivity($fsId) ? Carbon::parse($this->foodsaver['last_activity'])->format('d.m.Y') : '',
+                'lastActivity' => $this->profilePermissions->maySeelastActivity($fsId) ? $formattedLastActivity : '',
                 'buddyCount' => $this->foodsaver['stat_buddycount'],
                 'name' => $this->foodsaver['name'],
                 'fsId' => $this->foodsaver['id'],